pyproject.toml
[project]
name = "task-messenger-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Danial Shaikenov"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiosmtplib (==4.0.0)",
    "alembic (==1.15.1)",
    "amqp (==5.3.1)",
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "async-timeout (==5.0.1)",
    "asyncpg (==0.30.0)",
    "bcrypt (==3.2.2)",
    "billiard (==4.2.1)",
    "black (==25.1.0)",
    "celery (==5.5.2)",
    "certifi (==2025.1.31)",
    "cffi (==1.17.1)",
    "click (==8.2.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.1)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.6)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "importlib-metadata (==4.13.0)",
    "iniconfig (==2.0.0)",
    "itsdangerous (==2.2.0)",
    "kombu (==5.5.2)",
    "mako (==1.3.9)",
    "markupsafe (==3.0.2)",
    "mccabe (==0.7.0)",
    "mypy-extensions (==1.1.0)",
    "packaging (==24.2)",
    "passlib (==1.7.4)",
    "pathspec (==0.12.1)",
    "platformdirs (==4.3.8)",
    "pluggy (==1.5.0)",
    "prompt-toolkit (==3.0.51)",
    "pyasn1 (==0.6.1)",
    "pycodestyle (==2.13.0)",
    "pycparser (==2.22)",
    "pydantic (==2.10.5)",
    "pydantic-settings (==2.7.1)",
    "pydantic-core (==2.27.2)",
    "pyflakes (==3.3.2)",
    "pyjwt (==2.10.1)",
    "pyparsing (==3.2.1)",
    "pytest (==8.3.5)",
    "pytest-asyncio (==0.25.3)",
    "pytest-dotenv (==0.5.2)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.0.1)",
    "python-jose (==3.5.0)",
    "python-multipart (==0.0.20)",
    "pytz (==2024.2)",
    "redis (==5.2.1)",
    "rsa (==4.9.1)",
    "setuptools (==78.1.0)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.37)",
    "starlette (==0.41.3)",
    "typing-extensions (==4.12.2)",
    "tzdata (==2025.2)",
    "uvicorn (==0.34.0)",
    "vine (==5.1.0)",
    "wcwidth (==0.2.13)",
    "zipp (==3.21.0)",
    "ruff (>=0.11.13,<0.12.0)",
    "poetry-core (>=2.0.0,<3.0.0)",
    "mypy (>=1.17.1,<2.0.0)",
    "types-python-jose (>=3.5.0.20250531,<4.0.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["B", "E", "F", "W"]
ignore = ["E203", "E501", "F401", "B008"]
exclude = [".venv", "__pycache__", ".pytest_cache", ".idea", ".git", "venv"]